fix prototypes/signature for setgroups, etc.

+++ include/unistd.h
@@ -154,7 +154,7 @@ int getpagesize(void);
 int sethostname(const char *, size_t);
 int usleep(unsigned);
 unsigned ualarm(unsigned, unsigned);
-int setgroups(int, const gid_t []);
+int setgroups(size_t, const gid_t []);
 int setresuid(uid_t, uid_t, uid_t);
 int setresgid(gid_t, gid_t, gid_t);
 #endif
