+++ src/linux/setgroups.c
@@ -1,7 +1,7 @@
 #include <unistd.h>
 #include "syscall.h"
 
-int setgroups(int count, const gid_t list[])
+int setgroups(size_t count, const gid_t list[])
 {
 	return syscall(SYS_setgroups, count, list);
 }
